fix: setup of HotFoodCabinet

+++ app/src/main/java/com/gemini/energy/wordDocGenerator/SorterForWordDocumentGenerator.kt
@@ -668,7 +668,7 @@ class SorterForWordDocumentGenerator {
                             "Hot Food Cabinet",
                             single.energyPowerChange(),
                             costElectricity,
-                            single.age,
+                            single.age.toDouble(),
                             materialCost,
                             implementationCost,
                             savings,
